Minor correction on doxygen pages styling.
There seems to be no way to configure Doxygen to generate sections to starts from H2 instead of H1. Since those pages are now "embedded" inside Jekyll template, we have now two kinds of H1 which is not so nice. This commit attempt to correct the cosmetic styling to demote doxygen's H1 to Jekyll's H2.